// src/app/edit/data-types/types.ts
export type AlbumDataProps = {
  webSizePx?: string;
  webPhotoSizePx?: string;
  layoutPage?: string;
  bookName?: string;
  images?: Array<{
    id: string;
    filename: string;
    mimeType: string;
    s3Url: string;
    previewUrl: string;
    imageUrl: string;
    height: number;
    width: number;
    metadata?: {
      captureDate?: string;
      eventGroup?: string;
      isCover?: boolean;
      textAnnotation?: TextAnnotation;
      rotation?: number | null;
      zoom?: number | null;
      caption?: string;
      locationName?: string | null;
      event_tags?: string[] | string;
      zoomPosition?: ZoomPosition | null;
    };
  }>;
  bookDesign: string | string[];
};

export interface ImageDataProps {
  id: string;
  filename: string;
  mimeType: string;
  previewUrl: string;
  imageUrl: string;
  height: number;
  width: number;
  metadata?: {
    captureDate?: string;
    eventGroup?: string;
    isCover?: boolean;
    textAnnotation?: TextAnnotation;
    rotation?: number | null;
    zoom?: number | null;
    caption?: string;
    locationName?: string | null;
    event_tags?: string[] | string;
    zoomPosition?: ZoomPosition | null;
  };
}